% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/augment_pubchem.R
\name{augment_pubchem}
\alias{augment_pubchem}
\title{Lookup data from PubChem}
\usage{
augment_pubchem(chemtable)
}
\arguments{
\item{chemtable}{the data frame of the data about the compounds}

\item{CAS}{Optional, a vector of CAS numbers}
}
\description{
Lookup data from PubChem
}
