% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_filters.R
\name{combine_filters}
\alias{combine_filters}
\title{Combine filters}
\usage{
combine_filters(chemtable)
}
\arguments{
\item{chemtable}{the data frame of the data about the compounds}
}
\description{
Combine filters
}
\examples{
data(GCMSfloral)
chemtable <- combine_filters(chemtable)
}
\keyword{filters}
