% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_ambient_date.R
\name{filter_ambient_date}
\alias{filter_ambient_date}
\title{Filter on ratios between samples and controls by date}
\usage{
filter_ambient_date(chemtable, sampletable, metadata, ratio = 3)
}
\arguments{
\item{chemtable}{the data frame of the data about the compounds}

\item{sampletable}{the wide data frame with samples in rows and compound names in columns, containing peak areas}

\item{metadata}{the data frame that contains meta data about the group, type, and other attributes of each sample}

\item{ratio}{the minimum ratio of mean areas from floral to the ambient sample on that collection date}
}
\description{
This function allows you to filter compounds based on a threshold ratio of a compound's mean area in floral samples vs. ambient samples on a particular date.
}
\examples{
data(GCMSfloral)
chemtable <- filter_ambient_date(chemtable, sampletable, metadata, ratio=10)
}
\keyword{ambient}
\keyword{filter}
\keyword{floral}
