% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_ambient_presence.R
\name{filter_ambient_presence}
\alias{filter_ambient_presence}
\title{Filter on occurence in controls}
\usage{
filter_ambient_presence(chemtable, sampletable, metadata)
}
\arguments{
\item{chemtable}{the data frame of the data about the compounds}

\item{sampletable}{the wide data frame with samples in rows and compound names in columns, containing peak areas}

\item{metadata}{the data frame that contains meta data about the group, type, and other attributes of each sample}
}
\description{
This function allows you to filter compounds based on whether they are present in the ambient controls.
}
\examples{
data(GCMSfloral)
chemtable <- filter_ambient_presence(chemtable, sampletable, metadata)
}
\keyword{ambient}
\keyword{filter}
\keyword{floral}
