% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_area.R
\name{filter_area}
\alias{filter_area}
\title{Filter on the maximum peak area of a compound}
\usage{
filter_area(chemtable, min_maximum)
}
\arguments{
\item{chemtable}{the data frame of the data about the compounds}

\item{min_maximum}{the minimum value for maximum peak area of a compound}
}
\description{
This function allows you to filter compounds that only appear in very small amounts.
}
\examples{
data(GCMSfloral)
chemtable <- filter_area(chemtable, min_maximum = 20000)
}
\keyword{area}
\keyword{filter}
