% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_count.R
\name{filter_count}
\alias{filter_count}
\title{Filter on compound counts}
\usage{
filter_count(chemtable, min_count = 2, group = FALSE)
}
\arguments{
\item{chemtable}{the data frame of the data about the compounds}

\item{min_count}{the minimum count value for inclusion}

\item{group}{boolean, whether to set thresholds within each grouping level, or look across all floral samples}
}
\description{
This function allows you to filter compounds are rare in the entire dataset or within certain groups.
}
\examples{
data(GCMSfloral)
chemtable <- filter_count(chemtable, min_count = 3, group = TRUE)
}
\keyword{count}
\keyword{filter}
