% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_freq.R
\name{filter_freq}
\alias{filter_freq}
\title{Filter on compound frequency}
\usage{
filter_freq(chemtable, min_freq = 0.1, group = FALSE)
}
\arguments{
\item{chemtable}{the data frame of the data about the compounds}

\item{min_freq}{the minimum frequency value for inclusion}

\item{group}{boolean, whether to set thresholds within each grouping level, or look across all floral samples}
}
\description{
This function allows you to filter compounds are rare in the entire dataset or within certain groups.
}
\examples{
data(GCMSfloral)
chemtable <- filter_freq(chemtable, min_count = 0.2, group = TRUE)
}
\keyword{filter}
\keyword{frequency}
