% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_match.R
\name{filter_match}
\alias{filter_match}
\title{Filter on the match between a compound and a library}
\usage{
filter_match(chemtable, min = 0.7)
}
\arguments{
\item{chemtable}{the data frame of the data about the compounds}

\item{min}{the minimum match score out of 1.}
}
\description{
This function allows you to filter compounds that do not meet a minimum average match score with the library.
}
\examples{
data(GCMSfloral)
chemtable <- filter_match(chemtable, 0.7)
}
\keyword{filter}
\keyword{library}
\keyword{match}
