% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_longdata.R
\name{load_longdata}
\alias{load_longdata}
\title{Load GC-MS data}
\usage{
load_longdata(data, sample, RT, name, area, match, maxmatch = 100)
}
\arguments{
\item{data}{the long format data frame}

\item{sample}{the string that species the sample ID}

\item{RT}{the string that specifies the peak area retention time}

\item{name}{the string that specifies the compound name}

\item{area}{the string that specifies the integrated peak area}

\item{match}{the numeric match score (dot product of unknown and database spectra)}

\item{maxmatch}{numerical value that specifies the maximum value possible for the match score (usually 100 or 1000)}
}
\value{
longdata, a data frame with the standard column names and types, with match score relative to 1
}
\description{
Loads long-format GC-MS data and standardizes the column names and match scores
}
\examples{
read.table("GCMS_output.csv", header = TRUE, sep = "\\t") \%>\%
load_longdata("Sample", "Time", "NIST_ID", "Area", "Score")
}
