% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_metadata.R
\name{load_metadata}
\alias{load_metadata}
\title{Load metadata}
\usage{
load_metadata(metadata, date = NULL, sample, group = NULL, type,
  amount = NULL)
}
\arguments{
\item{metadata}{data frame of sample metadata}

\item{date}{the text date of sampling}

\item{sample}{the string that specifies the sample ID}

\item{group}{a vector of columns that specifies species or treatment groups}

\item{type}{the factor that specifies what type of sample it is (e.g. floral, ambient control, etc.). Levels should be c("floral", "ambient", "leaf")}

\item{amount}{the string that specifies the amount used for standardization (e.g. flower number or mass)}
}
\value{
metadata, a data frame with the standard column names and types
}
\description{
Loads metadata about the samples
}
\examples{
data(GCMSfloral)
metadata <- load_metadata(GCMS_metadata, "SampleDate", "Filename", c("Cross", "Time"), "Type", "Flrs")
}
