% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_chemtable.R
\name{make_chemtable}
\alias{make_chemtable}
\title{Make chemical table}
\usage{
make_chemtable(longdata, metadata)
}
\arguments{
\item{longdata}{the long format data frame}

\item{metadata}{the sample metadata}
}
\value{
chemtable, a data frame of the compounds, with the following columns:
compound: compound name
RT: mean RT
RT.var: variance of RT
match: mean match
match.var: variance of match
max: maximum area
mean: mean area
count: number of occurrences across all samples
TODO count.X: multiple columns, number of occurrences within types and treatment groups
freq: proportion of occurrences across all samples
TODO freq.X: multiple columns, proportion of occurrences within types and treatment groups
}
\description{
Make chemical table
}
\examples{
chemtable <- make_chemtable(longdata, metadata)
}
