% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_filters.R
\name{plot_filters}
\alias{plot_filters}
\title{Diagnostic plot of compound filtering}
\usage{
plot_filters(
  chemtable,
  option = "rarity",
  yrange = 3,
  fontsize = 3,
  pointsize = 10,
  alpha_excluded = 0
)
}
\arguments{
\item{chemtable}{the data frame of the data about the compounds}

\item{yrange}{the number of orders of magnitude the log10 y-axis will span (for method "ambient") or the negative left limit of the  log2 x-axis (for method "volcano"), NA means no limit}

\item{fontsize}{size of the compound labels}

\item{pointsize}{size of the largest point}

\item{alpha_excluded}{transparency of text for compounds that passed some test but were excluded from the final filter}

\item{method}{the type of plot to produce, "rarity" or "ambient", see Details}
}
\description{
This function creates several plots that visualizes compounds and filtering criteria. Requires filter_area and filter_freq to be run first, and optionally filter_ambient_ratio. It requires ggplot2.
}
\details{
The following plotting options are available. The color of points is set based on whether the compound met
the final filtering criteria (filter_final column), or passed some test. See the pipeline vignette for details.
\describe{
  \item{rarity}{Plots compound rarity (x-axis),
maximum peak area across samples (y-axis), mean nonzero area peak area (size), and
amounts relative to ambient controls (labels, if filter_ambient_ratio is run first)}
  \item{ambient}{Plots mean peak area in ambient controls (x-axis) versus in floral samples (y-axis),
  frequency in floral samples (size), and shows the cutoff used for filter_ambient_ratio (dotted line)}
  \item{volcano}{Plots fold change (x-axis) and (adjusted) p-value (y-axis) comparing floral to ambient samples,
  frequency in floral samples (size), and shows the cutoff used for filter_ambient_ratio and alpha (dotted lines)}
  \item{prop}{Plots the proportion of chemicals that passed each test}
}
}
\examples{
data(GCMSfloral)
plot_filters(chemtable, option="rarity")
plot_filters(chemtable, option="ambient")
}
\keyword{filter}
\keyword{plot}
