% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prune_sampletable_dates.R
\name{prune_sampletable_dates}
\alias{prune_sampletable_dates}
\title{Prune sampletable by date
Zeros out cells in the sample table that did not pass the ambient date filter.}
\usage{
prune_sampletable_dates(sampletable, chemtable, metadata, datefilter)
}
\arguments{
\item{chemtable}{the data frame of the data about the compounds}
}
\description{
Prune sampletable by date
Zeros out cells in the sample table that did not pass the ambient date filter.
}
\examples{
data(GCMSfloral)
chemtable <- prune_sampletable_dates(ampletable, chemtable, metadata, datefilter)
}
\keyword{filters}
