% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prune_sampletable_leaf.R
\name{prune_sampletable_leaf}
\alias{prune_sampletable_leaf}
\title{Prune sampletable by leaf
Zeros out cells in the sample table that did not pass the ambient leaf filter.}
\usage{
prune_sampletable_leaf(sampletable, chemtable, metadata, leafgroup, leaffilter)
}
\arguments{
\item{chemtable}{the data frame of the data about the compounds}
}
\description{
Prune sampletable by leaf
Zeros out cells in the sample table that did not pass the ambient leaf filter.
}
\examples{
data(GCMSfloral)
chemtable <- prune_sampletable_leaf(ampletable, chemtable, metadata, leafgroup, leaffilter)
}
\keyword{filters}
