% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardize_finaltable.R
\name{standardize_finaltable}
\alias{standardize_finaltable}
\title{Standardize finaltable by amount or row total}
\usage{
standardize_finaltable(finaltable, metadata, percent = FALSE)
}
\arguments{
\item{finaltable}{the wide data frame with samples in rows and compound names in columns, containing peak areas for floral samples only}

\item{metadata}{the data frame that contains meta data about the group, type, and other attributes of each sample}

\item{percent}{if TRUE, divide by row sums instead of amount}
}
\value{
finaltable
}
\description{
Divides peak areas by the amount (biomass, counts) of each floral sample in metadata, or by the sample sum to create percentages.
}
\examples{
data(GCMSfloral)
standardize_finaltable(finaltable, metadata, percent = FALSE)
}
