% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_ambient_ttest.R
\name{filter_ambient_ttest}
\alias{filter_ambient_ttest}
\title{Filter on significant differences from controls}
\usage{
filter_ambient_ttest(
  chemtable,
  sampletable,
  metadata,
  alpha,
  adjust = "none",
  compare = "ambient"
)
}
\arguments{
\item{chemtable}{the data frame of the data about the compounds}

\item{sampletable}{the wide data frame with samples in rows and compound names in columns, containing peak areas}

\item{metadata}{the data frame that contains meta data about the group, type, and other attributes of each sample}

\item{alpha}{maximum acceptable p-value Defaults to XXX}

\item{adjust}{"fdr" for False Discovery Rate correction, "none" for no correction, or any other method of p.adjust}

\item{compare}{the type of sample to compare against - ambient or blank}
}
\description{
This function allows you to filter compounds based on whether the quantity of a compound in ambient or blank samples vs. floral samples is significantly different.
}
\examples{
filter_ambient_ttest(chemtable,sampletable,metadata,0.05)
}
\keyword{ambient}
\keyword{filter}
\keyword{floral}
