% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_RT.R
\name{filter_RT}
\alias{filter_RT}
\title{Filter on retention times}
\usage{
filter_RT(chemtable, RT_min, RT_max)
}
\arguments{
\item{chemtable}{the data frame of the data about the compounds}

\item{RT_min}{the RT_minimum retention time}

\item{RT_max}{the RT_maximum retention time}
}
\description{
This function allows you to filter out compounds that elute before a RT_minimum retention time and/or after a RT_maximum retention time.
}
\examples{
data(GCMSfloral)
chemtable <- filter_RT(chemtable, 4, 15)
}
\keyword{filter}
\keyword{retention}
\keyword{time}
