% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_RT.R
\name{filter_RT}
\alias{filter_RT}
\title{A function to filter on retention times}
\usage{
filter_RT(chemtable, min, max)
}
\arguments{
\item{chemtable}{the data frame of the data about the compounds
min the minimum retention time Defaults to xxx
max the maximum retention time Defaults to xxx}
}
\description{
This function allows you to filter compounds that elute before a minimum retention time and/or after a maximum retention time.
}
\examples{
filter_RT()
}
\keyword{filter}
\keyword{retention}
\keyword{time}
