% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_chemtable.R
\name{make_chemtable}
\alias{make_chemtable}
\title{Make chemical table}
\usage{
make_chemtable(longdata, metadata)
}
\arguments{
\item{longdata}{the long format data frame}

\item{metadata}{the sample metadata}
}
\value{
chemtable, a data frame of the compound
}
\description{
Summarize information about each chemical, including retention time and frequency of occurence by groups. Prerequisite for the filter* functions.
}
\details{
Generates chemtable, a data frame of the compounds, with the following columns:
\describe{
\item{compound}{compound name}
\item{RT}{mean RT}
\item{RT.var}{variance of RT}
\item{match}{mean match}
\item{match.var}{variance of match}
\item{max.area}{maximum area in floral samples}
\item{mean.area}{mean area in floral samples, including zeros}
\item{count.X}{multiple columns, number of occurrences within types and treatment groups}
\item{freq.X}{multiple columns, proportion of occurrences within types and treatment groups}
}
}
\examples{
data(GCMSfloral)
\dontrun{chemtable <- make_chemtable(longdata, metadata)}
}
