% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_chemtable.R
\name{make_chemtable}
\alias{make_chemtable}
\title{Make chemical table}
\usage{
make_chemtable(gcdata, metadata, datatype = "long")
}
\arguments{
\item{gcdata}{peaks integrated from GC data, either in long format (peaks in rows with columns for area, match score, and retention time, as generated by `load_longdata()`) or if detailed peak information is not available, in wide format (table of peak areas with samples in rows and compounds in columns, as generated by `make_sampletable()`).}

\item{metadata}{the sample metadata, in the same row order as gcdata if datatype is 'wide'}

\item{datatype}{the format of gcdata, either 'long' (the default) or 'wide'}
}
\value{
chemtable, a data frame of the compound
}
\description{
Summarize information about each chemical, including retention time and frequency of occurence by groups. Prerequisite for the filter* functions.
}
\details{
Generates chemtable, a data frame of the compounds, with the following columns:
\describe{
\item{compound}{compound name}
\item{RT}{mean RT}
\item{RT.var}{variance of RT}
\item{match}{mean match}
\item{match.var}{variance of match}
\item{max.floral}{maximum area in floral samples}
\item{mean.floral}{mean area in floral samples, including zeros}
\item{mean.nonzero.floral}{mean area in floral samples, excluding zeros}
\item{median.floral}{median area in floral samples, including zeros}
\item{count.X}{multiple columns, number of occurrences within types and treatment groups}
\item{freq.X}{multiple columns, proportion of occurrences within types and treatment groups}
}
}
\examples{
data(GCMSfloral)
\dontrun{chemtable <- make_chemtable(longdata, metadata)}
}
