% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_sampletable.R
\name{make_sampletable}
\alias{make_sampletable}
\title{Make sample table}
\usage{
make_sampletable(longdata)
}
\arguments{
\item{longdata}{the long format data frame}
}
\value{
sampletable, a wide data frame with samples in rows and compound names in columns, containing peak areas
}
\description{
Converts long-format GC-MS data to a wide data frame with samples in rows and compound names in columns, containing peak areas
}
\examples{
sampletable <- make_sampletable(longdata)
}
