% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_filters.R
\name{plot_filters}
\alias{plot_filters}
\title{Diagnostic plot of compound filtering}
\usage{
plot_filters(chemtable, yrange = 3, fontsize = 3, pointsize = 15)
}
\arguments{
\item{chemtable}{the data frame of the data about the compounds}

\item{yrange}{the number of order of magnitudes the log10 y-axis will span}

\item{fontsize}{size of the compound labels}

\item{pointsize}{size of the largest point}
}
\description{
This function creates a plot that visualizes compound rarity (x-axis),
maximum peak area across samples (y-axis), mean nonzero area peak area (size),
amounts relative to ambient controls (labels), and whether the compound met
the final filtering critera (filter_final column), or the area and frquency filters.
Requires filter_area and gilter_freq to be run first.
}
\examples{
data(GCMSfloral)
chemtable <- make_chemtable(longdata, metadata)
chemtable <- filter_area(chemtable, min_maximum = 20000)
chemtable <- filter_freq(chemtable, 0.2)
chemtable <- combine_filters(chemtable)
plot_filters(chemtable)
}
\keyword{filter}
\keyword{plot}
